Shorten the urls during at rendering
